Lots changed
Have not been disciplined about commiting at small steps, so instead of
a message on what's changed, this is a current state message.
Got static opcode working.
EX is coming along. Many R-type instructions implemented, but not tested
EX tests are in progress. Implementation got a little ahead of tests.